Add GroupCache optimized for removing related keys
This commit adds GroupCache that provides the same functionality
as Cache, as well as RemoveGroup() and RemoveGroups() features.

RemoveGroup() and RemoveGroups() performance is linear to the
number of cache items to remove, regardless of current cache size.

As a tradeoff, GroupCache adds a small memory overhead compared
to regular Cache, and the memory overhead is linear to
cache key sizes.